a2a/pyproject.toml
[project]
name = "a2a"
version = "0.1.0"
description = "A2A submodule for Restate and GenAI integrations"
requires-python = ">=3.12"

dependencies = [
    "httpx>=0.28.1",
    "hypercorn",
    "restate-sdk[serde]>=0.11.0",
    "pydantic>=2.11.9",
    "openai",
    "fastapi[standard]",
    "python-dotenv>=1.1.0",
    "openai-agents==0.3.2",
    "a2a-sdk>=0.3.0",
    "click>=8.1.8",
    "google-adk>=1.8.0",
    "google-genai>=1.27.0",
    "python-dotenv>=1.1.0",
    "litellm",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.18.2",
]
